requirements.txt
absl-py==0.7.1
astor==0.7.1
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==2.2.4
mock==2.0.0
numpy==1.16.3
pbr==5.2.0
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
scipy==1.2.1
six==1.12.0
subprocess32==3.5.3
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-probability==0.6.0
termcolor==1.1.0
transforms3d==0.3.1
Werkzeug==0.15.2
